backend/requirements.txt
django==4.0
djangorestframework==3.13.0
djangorestframework-camel-case==1.4.2
django-extensions==3.2.1
drf-extensions==0.7.1
django-cors-headers
django-simpleui
django-modeltranslation
drf-yasg2
django-filter
django-redis
redis
django-redis-cache
django-celery-beat
celery==5.3.4
django-celery-results
requests
sentry-sdk
django-ckeditor
python-dotenv
pycryptodome
mysqlclient
azure-communication-email
Pillow==9.5.0
channels
daphne
azure-storage-blob
django-storages[azure]
mongoengine
azure-servicebus
locust
django-import-export
discord.py
bs4
tweepy
google_auth_oauthlib
google-api-python-client
moviepy
django-ckeditor
kombu==5.3.2
numpy
python-telegram-bot
openpyxl
pandas
web3==6.10.0
